Add files via upload
This project is a clone website of the popular music application of SPOTIFY. This project gives us a demo of the visuals, graphics, and the system used by Spotify. The main purpose behind developing this Mini Project was to understand the working of the existential systems and the loopholes(if any) and also to check if there are any upgradations required in the existing system.